- restrict mapboxSdkService dependency from `libnavigation-base` and `libdirections-common` module; - added mapboxSdkService dependency to `libdirections-offboard` dependency
