+++ libdirections-offboard/src/main/java/com/mapbox/navigation/route/offboard/extension/Mappers.kt
@@ -1,77 +1,20 @@
 package com.mapbox.navigation.route.offboard.extension
 
-import com.mapbox.api.directions.v5.DirectionsCriteria
-import com.mapbox.api.directions.v5.WalkingOptions
+import android.location.Location
 import com.mapbox.api.directions.v5.models.DirectionsRoute
-import com.mapbox.api.directions.v5.models.RouteOptions
+import com.mapbox.geojson.Point
 import com.mapbox.navigation.base.route.model.Route
 import com.mapbox.navigation.base.route.model.RouteLegsNavigation
-import com.mapbox.navigation.base.route.model.RouteOptionsNavigation
-import com.mapbox.navigation.base.route.model.WalkingOptionsNavigation
-import java.util.Locale
+
+fun Location.toPoint(): Point = Point.fromLngLat(this.longitude, this.latitude)
 
 fun DirectionsRoute.mapToRoute() = Route(
     routeIndex = routeIndex(),
-    distance = distance()!!,
-    duration = duration()?.toLong()!!,
+    distance = distance(),
+    duration = duration()?.toLong(),
     geometry = geometry(),
     weight = weight(),
     weightName = weightName(),
-    legs = legs()?.let { RouteLegsNavigation(it) },
-    routeOptions = routeOptions()?.mapToRouteOptionsNavigation(),
-    voiceLanguage = voiceLanguage()
+    voiceLanguage = voiceLanguage(),
+    legs = legs()?.let { RouteLegsNavigation(it) }
 )
-
-fun RouteOptions.mapToRouteOptionsNavigation(): RouteOptionsNavigation {
-    val routeOptionsNavigationBuilder = RouteOptionsNavigation
-        .builder()
-        .baseUrl(baseUrl())
-        .user(user())
-        .profile(profile())
-        .origin(coordinates().first())
-        .destination(coordinates().last())
-    coordinates().drop(1).dropLast(1)
-        .forEach { routeOptionsNavigationBuilder.addWaypoint(it) }
-    return routeOptionsNavigationBuilder
-        .alternatives(alternatives() ?: true)
-        .language(language() ?: Locale.getDefault().language)
-        .radiuses(radiuses() ?: "")
-        .bearings(bearings() ?: "")
-        .continueStraight(continueStraight() ?: true)
-        .roundaboutExits(roundaboutExits() ?: true)
-        .geometries(geometries() ?: DirectionsCriteria.GEOMETRY_POLYLINE6)
-        .overview(overview() ?: DirectionsCriteria.OVERVIEW_FULL)
-        .steps(steps() ?: true)
-        .annotations(
-            annotations() ?: arrayOf(
-                DirectionsCriteria.ANNOTATION_CONGESTION,
-                DirectionsCriteria.ANNOTATION_DISTANCE
-            ).joinToString(separator = ",")
-        )
-        .voiceInstructions(voiceInstructions() ?: true)
-        .bannerInstructions(bannerInstructions() ?: true)
-        .voiceUnits(voiceUnits() ?: "")
-        .accessToken(accessToken())
-        .requestUuid(requestUuid())
-        .exclude(exclude() ?: "")
-        .approaches(approaches() ?: "")
-        .waypointIndices(waypointIndices() ?: "")
-        .waypointNames(waypointNames() ?: "")
-        .waypointTargets(waypointTargets() ?: "")
-        .walkingOptions(walkingOptions()?.mapToWalkingOptionsNavigation()!!)
-        .build()
-}
-
-fun WalkingOptionsNavigation.mapToWalkingOptions(): WalkingOptions = WalkingOptions
-    .builder()
-    .walkingSpeed(walkingSpeed)
-    .walkwayBias(walkwayBias)
-    .alleyBias(alleyBias)
-    .build()
-
-fun WalkingOptions.mapToWalkingOptionsNavigation(): WalkingOptionsNavigation =
-    WalkingOptionsNavigation(
-        walkingSpeed = walkingSpeed(),
-        walkwayBias = walkwayBias(),
-        alleyBias = alleyBias()
-    )
