+++ libdirections-offboard/src/main/java/com/mapbox/navigation/route/offboard/router/NavigationRoute.kt
@@ -1,16 +1,18 @@
 package com.mapbox.navigation.route.offboard.router
 
 import android.content.Context
+import androidx.annotation.FloatRange
+import androidx.annotation.IntRange
 import com.mapbox.api.directions.v5.DirectionsCriteria
 import com.mapbox.api.directions.v5.MapboxDirections
 import com.mapbox.api.directions.v5.models.DirectionsResponse
 import com.mapbox.api.directions.v5.models.RouteOptions
+import com.mapbox.core.exceptions.ServicesException
+import com.mapbox.core.utils.TextUtils
 import com.mapbox.geojson.Point
-import com.mapbox.navigation.base.route.model.RouteOptionsNavigation
-import com.mapbox.navigation.base.route.model.RoutePointNavigation
 import com.mapbox.navigation.route.offboard.extension.getUnitTypeForLocale
-import com.mapbox.navigation.route.offboard.extension.mapToWalkingOptions
 import com.mapbox.navigation.utils.extensions.inferDeviceLocale
+import com.mapbox.services.android.navigation.v5.utils.extensions.getUnitTypeForLocale
 import java.util.Locale
 import okhttp3.EventListener
 import okhttp3.Interceptor
@@ -26,16 +28,17 @@ import retrofit2.Callback
  *
  * Developer Note: MapboxDirections cannot be directly extended since it is an AutoValue class.
  *
- * 1.0
+ * 0.5.0
  */
 
 class NavigationRoute
-constructor(
+internal constructor(
     private val mapboxDirections: MapboxDirections
 ) {
 
     companion object {
-        private val EVENT_LISTENER = NavigationRouteEventListener()
+        private val EVENT_LISTENER =
+            NavigationRouteEventListener()
 
         /**
          * Build a new [NavigationRoute] object with the proper navigation parameters already setup.
@@ -46,21 +49,14 @@ constructor(
         @JvmStatic
         fun builder(context: Context): Builder =
             Builder()
-                .profile(DirectionsCriteria.PROFILE_DRIVING_TRAFFIC)
-                .language(context)
-                .continueStraight(true)
-                .roundaboutExits(true)
-                .geometries(DirectionsCriteria.GEOMETRY_POLYLINE6)
-                .overview(DirectionsCriteria.OVERVIEW_FULL)
-                .steps(true)
                 .annotations(
                     DirectionsCriteria.ANNOTATION_CONGESTION,
                     DirectionsCriteria.ANNOTATION_DISTANCE
                 )
-                .voiceInstructions(true)
-                .bannerInstructions(true)
-                .enableRefresh(true)
+                .language(context)
                 .voiceUnits(context)
+                .profile(DirectionsCriteria.PROFILE_DRIVING_TRAFFIC)
+                .continueStraight(true)
     }
 
     /**
@@ -81,7 +77,12 @@ constructor(
      * @since 0.5.0
      */
     fun getRoute(callback: Callback<DirectionsResponse>) {
-        mapboxDirections.enqueueCall(callback)
+        mapboxDirections.enqueueCall(
+            NavigationRouteCallback(
+                EVENT_LISTENER,
+                callback
+            )
+        )
     }
 
     /**
@@ -109,9 +110,9 @@ constructor(
      */
     class Builder internal constructor(private val directionsBuilder: MapboxDirections.Builder) {
         private val eventListener: NavigationRouteEventListener
-        private var origin: RoutePointNavigation? = null
-        private var destination: RoutePointNavigation? = null
-        private val waypoints = ArrayList<RoutePointNavigation>()
+        private var origin: NavigationRouteWaypoint? = null
+        private var destination: NavigationRouteWaypoint? = null
+        private val waypoints = ArrayList<NavigationRouteWaypoint>()
         private val SEMICOLON = ";"
         private val COMMA = ","
 
@@ -121,7 +122,22 @@ constructor(
         constructor() : this(MapboxDirections.builder()) {}
 
         init {
-            this.eventListener = EVENT_LISTENER
+            this.eventListener =
+                EVENT_LISTENER
+        }
+
+        /**
+         * The username for the account that the directions engine runs on. In most cases, this should
+         * always remain the default value of [DirectionsCriteria.PROFILE_DEFAULT_USER].
+         *
+         * @param user a non-null string which will replace the default user used in the directions
+         * request
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun user(user: String): Builder {
+            directionsBuilder.user(user)
+            return this
         }
 
         /**
@@ -133,23 +149,192 @@ constructor(
          * @return this builder for chaining options together
          * @since 0.5.0
          */
-        internal fun profile(@DirectionsCriteria.ProfileCriteria profile: String): Builder {
+        fun profile(@DirectionsCriteria.ProfileCriteria profile: String): Builder {
             directionsBuilder.profile(profile)
             return this
         }
 
         /**
-         * Sets allowed direction of travel when departing intermediate waypoints. If true the route
-         * will continue in the same direction of travel. If false the route may continue in the
-         * opposite direction of travel. API defaults to true for
-         * [DirectionsCriteria.PROFILE_DRIVING] and false for
-         * [DirectionsCriteria.PROFILE_WALKING] and [DirectionsCriteria.PROFILE_CYCLING].
+         * This sets the starting point on the map where the route will begin. It is one of the
+         * required parameters which must be set for a successful directions response.
          *
-         * @param continueStraight boolean true if you want to always continue straight, else false.
+         * @param origin a GeoJson [Point] object representing the starting location for the route
          * @return this builder for chaining options together
+         * @since 0.5.0
          */
-        internal fun continueStraight(continueStraight: Boolean): Builder {
-            directionsBuilder.continueStraight(continueStraight)
+        fun origin(origin: Point): Builder {
+            this.origin = NavigationRouteWaypoint(
+                origin,
+                null,
+                null
+            )
+            return this
+        }
+
+        /**
+         * This sets the starting point on the map where the route will begin. It is one of the
+         * required parameters which must be set for a successful directions response.
+         *
+         * @param origin a GeoJson [Point] object representing the starting location for the
+         * route
+         * @param angle double value used for setting the corresponding coordinate's angle of travel
+         * when determining the route
+         * @param tolerance the deviation the bearing angle can vary while determining the route,
+         * recommended to be either 45 or 90 degree tolerance
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun origin(
+            origin: Point,
+            angle: Double?,
+            tolerance: Double?
+        ): Builder {
+            this.origin =
+                NavigationRouteWaypoint(
+                    origin,
+                    angle,
+                    tolerance
+                )
+            return this
+        }
+
+        /**
+         * This sets the ending point on the map where the route will end. It is one of the required
+         * parameters which must be set for a successful directions response.
+         *
+         * @param destination a GeoJson [Point] object representing the starting location for the
+         * route
+         * @return this builder for chaining options together
+         * @since 0.50
+         */
+        fun destination(destination: Point): Builder {
+            this.destination =
+                NavigationRouteWaypoint(
+                    destination,
+                    null,
+                    null
+                )
+            return this
+        }
+
+        /**
+         * This sets the ending point on the map where the route will end. It is one of the required
+         * parameters which must be set for a successful directions response.
+         *
+         * @param destination a GeoJson [Point] object representing the starting location for the
+         * route
+         * @param angle double value used for setting the corresponding coordinate's angle of travel
+         * when determining the route
+         * @param tolerance the deviation the bearing angle can vary while determining the route,
+         * recommended to be either 45 or 90 degree tolerance
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun destination(
+            destination: Point,
+            angle: Double?,
+            tolerance: Double?
+        ): Builder {
+            this.destination =
+                NavigationRouteWaypoint(
+                    destination,
+                    angle,
+                    tolerance
+                )
+            return this
+        }
+
+        /**
+         * This can be used to set up to 23 additional in-between points which will act as pit-stops
+         * along the users route. Note that if you are using the
+         * [DirectionsCriteria.PROFILE_DRIVING_TRAFFIC] that the max number of waypoints allowed
+         * in the request is currently limited to 1.
+         *
+         * @param waypoint a [Point] which represents the pit-stop or waypoint where you'd like
+         * one of the [com.mapbox.api.directions.v5.models.RouteLeg] to
+         * navigate the user to
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun addWaypoint(waypoint: Point): Builder {
+            this.waypoints.add(
+                NavigationRouteWaypoint(
+                    waypoint,
+                    null,
+                    null
+                )
+            )
+            return this
+        }
+
+        /**
+         * This can be used to set up to 23 additional in-between points which will act as pit-stops
+         * along the users route.
+         *
+         *
+         * Note that if you are using the
+         * [DirectionsCriteria.PROFILE_DRIVING_TRAFFIC] that the max number of waypoints allowed
+         * in the request is currently limited to 1.
+         *
+         *
+         * These waypoints are added to the request in the order you add them to the builder with this method.
+         *
+         * @param waypoint a [Point] which represents the pit-stop or waypoint where you'd like
+         * one of the [com.mapbox.api.directions.v5.models.RouteLeg] to
+         * navigate the user to
+         * @param angle double value used for setting the corresponding coordinate's angle of travel
+         * when determining the route
+         * @param tolerance the deviation the bearing angle can vary while determining the route,
+         * recommended to be either 45 or 90 degree tolerance
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun addWaypoint(
+            waypoint: Point,
+            angle: Double?,
+            tolerance: Double?
+        ): Builder {
+            this.waypoints.add(
+                NavigationRouteWaypoint(
+                    waypoint,
+                    angle,
+                    tolerance
+                )
+            )
+            return this
+        }
+
+        /**
+         * Optionally set whether to try to return alternative routes. An alternative is classified as a
+         * route that is significantly different then the fastest route, but also still reasonably fast.
+         * Not in all circumstances such a route exists. At the moment at most one alternative can be
+         * returned.
+         *
+         * @param alternatives true if you'd like to receive an alternative route, otherwise false or
+         * null to use the APIs default value
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun alternatives(alternatives: Boolean?): Builder {
+            directionsBuilder.alternatives(alternatives)
+            return this
+        }
+
+        /**
+         * Set the instruction language for the directions request, the default is english. Only a
+         * select number of languages are currently supported, reference the table provided in the see
+         * link below.
+         *
+         * @param language a Locale representing the language you'd like the instructions to be
+         * written in when returned
+         * @return this builder for chaining options together
+         * @see [Supported
+         * Languages](https://www.mapbox.com/api-documentation/.instructions-languages)
+         *
+         * @since 0.5.0
+         */
+        fun language(language: Locale): Builder {
+            directionsBuilder.language(language)
             return this
         }
 
@@ -158,43 +343,126 @@ constructor(
             return this
         }
 
-        internal fun roundaboutExits(roundaboutExits: Boolean): Builder {
-            directionsBuilder.roundaboutExits(roundaboutExits)
-            return this
-        }
-
-        internal fun geometries(@DirectionsCriteria.GeometriesCriteria geometry: String): Builder {
-            directionsBuilder.geometries(geometry)
-            return this
-        }
-
-        internal fun overview(@DirectionsCriteria.OverviewCriteria overview: String): Builder {
-            directionsBuilder.overview(overview)
-            return this
-        }
-
-        internal fun steps(steps: Boolean): Builder {
-            directionsBuilder.steps(steps)
-            return this
-        }
-
-        internal fun annotations(@DirectionsCriteria.AnnotationCriteria vararg annotations: String?): Builder {
+        /**
+         * Whether or not to return additional metadata along the route. Possible values are:
+         * [DirectionsCriteria.ANNOTATION_DISTANCE],
+         * [DirectionsCriteria.ANNOTATION_DURATION],
+         * [DirectionsCriteria.ANNOTATION_DURATION] and
+         * [DirectionsCriteria.ANNOTATION_CONGESTION]. Several annotation can be used by
+         * separating them with `,`.
+         *
+         *
+         * If left alone, this will automatically set Congestion to enabled
+         *
+         *
+         * @param annotations string referencing one of the annotation direction criteria's. The strings
+         * restricted to one or multiple values inside the [AnnotationCriteria]
+         * or null which will result in no annotations being used
+         * @return this builder for chaining options together
+         * @see [RouteLeg object
+         * documentation](https://www.mapbox.com/api-documentation/.routeleg-object)
+         *
+         * @since 0.5.0
+         */
+        fun annotations(@DirectionsCriteria.AnnotationCriteria vararg annotations: String?): Builder {
             directionsBuilder.annotations(*annotations)
             return this
         }
 
-        internal fun voiceInstructions(voiceInstructions: Boolean): Builder {
-            directionsBuilder.voiceInstructions(voiceInstructions)
+        /**
+         * Optionally, Use to filter the road segment the waypoint will be placed on by direction and
+         * dictates the angle of approach. This option should always be used in conjunction with the
+         * [.radiuses] parameter.
+         *
+         *
+         * The parameter takes two values per waypoint: the first is an angle clockwise from true north
+         * between 0 and 360. The second is the range of degrees the angle can deviate by. We recommend
+         * a value of 45 degrees or 90 degrees for the range, as bearing measurements tend to be
+         * inaccurate. This is useful for making sure we reroute vehicles on new routes that continue
+         * traveling in their current direction. A request that does this would provide bearing and
+         * radius values for the first waypoint and leave the remaining values empty. If provided, the
+         * list of bearings must be the same length as the list of waypoints, but you can skip a
+         * coordinate and show its position by passing in null value for both the angle and tolerance
+         * values.
+         *
+         *
+         * Each bearing value gets associated with the same order which coordinates are arranged in this
+         * builder. For example, the first bearing added in this builder will be associated with the
+         * origin `Point`, the nth bearing being associated with the nth waypoint added (if added)
+         * and the last bearing being added will be associated with the destination.
+         *
+         *
+         * If given the chance, you should pass in the bearing information at the same time the point is
+         * passed in as a waypoint, this way it is ensured the value is matched up correctly with the
+         * coordinate.
+         *
+         * @param angle double value used for setting the corresponding coordinate's angle of travel
+         * when determining the route
+         * @param tolerance the deviation the bearing angle can vary while determining the route,
+         * recommended to be either 45 or 90 degree tolerance
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        @Deprecated(
+            "use the bearing paired with {@link Builder#origin(Point, Double, Double)},\n" +
+                "      {@link Builder#destination(Point, Double, Double)},\n" +
+                "      or {@link Builder#addWaypoint(Point, Double, Double)} instead."
+        )
+        fun addBearing(
+            @FloatRange(from = 0.0, to = 360.0) angle: Double?,
+            @FloatRange(from = 0.0, to = 360.0) tolerance: Double?
+        ): Builder {
+            directionsBuilder.addBearing(angle, tolerance)
             return this
         }
 
-        internal fun bannerInstructions(bannerInstructions: Boolean): Builder {
-            directionsBuilder.bannerInstructions(bannerInstructions)
+        /**
+         * Optionally, set the maximum distance in meters that each coordinate is allowed to move when
+         * snapped to a nearby road segment. There must be as many radiuses as there are coordinates in
+         * the request. Values can be any number greater than 0 or they can be unlimited simply by
+         * passing [Double.POSITIVE_INFINITY].
+         *
+         *
+         * If no routable road is found within the radius, a `NoSegment` error is returned.
+         *
+         *
+         * @param radiuses double array containing the radiuses defined in unit meters.
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun radiuses(@FloatRange(from = 0.0) vararg radiuses: Double): Builder {
+            directionsBuilder.radiuses(*radiuses)
+            return this
+        }
+
+        /**
+         * Change the units used for voice announcements, this does not change the units provided in
+         * other fields outside of the [com.mapbox.api.directions.v5.models.VoiceInstructions]
+         * object.
+         *
+         * @param voiceUnits one of the values found inside the [VoiceUnitCriteria]
+         * @return this builder for chaining options together
+         * @since 0.8.0
+         */
+        fun voiceUnits(@DirectionsCriteria.VoiceUnitCriteria voiceUnits: String): Builder {
+            directionsBuilder.voiceUnits(voiceUnits)
             return this
         }
 
         internal fun voiceUnits(context: Context): Builder {
             directionsBuilder.voiceUnits(context.inferDeviceLocale().getUnitTypeForLocale())
+            return this
+        }
+
+        /**
+         * Exclude specific road classes such as highways, tolls, and more.
+         *
+         * @param exclude one of the values found inside the [ExcludeCriteria]
+         * @return this builder for chaining options together
+         * @since 0.8.0
+         */
+        fun exclude(@DirectionsCriteria.ExcludeCriteria exclude: String?): Builder {
+            directionsBuilder.exclude(exclude)
             return this
         }
 
@@ -207,6 +475,109 @@ constructor(
          */
         fun clientAppName(clientAppName: String): Builder {
             directionsBuilder.clientAppName(clientAppName)
+            return this
+        }
+
+        /**
+         * Required to call when this is being built. If no access token provided,
+         * [ServicesException] will be thrown.
+         *
+         * @param accessToken Mapbox access token, You must have a Mapbox account inorder to use
+         * the Optimization API
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun accessToken(accessToken: String): Builder {
+            directionsBuilder.accessToken(accessToken)
+            return this
+        }
+
+        /**
+         * Optionally change the APIs base URL to something other then the default Mapbox one.
+         *
+         * @param baseUrl base url used as end point
+         * @return this builder for chaining options together
+         * @since 0.5.0
+         */
+        fun baseUrl(baseUrl: String): Builder {
+            directionsBuilder.baseUrl(baseUrl)
+            return this
+        }
+
+        /**
+         * Indicates from which side of the road to approach a waypoint.
+         * Accepts <tt>unrestricted</tt> (default), <tt>curb</tt> or <tt>null</tt>.
+         * If set to <tt>unrestricted</tt>, the route can approach waypoints
+         * from either side of the road. If set to <tt>curb</tt>, the route will be returned
+         * so that on arrival, the waypoint will be found on the side that corresponds with the
+         * <tt>driving_side</tt> of the region in which the returned route is located.
+         * If provided, the list of approaches must be the same length as the list of waypoints.
+         *
+         * @param approaches null if you'd like the default approaches,
+         * else one of the options found in
+         * [com.mapbox.api.directions.v5.DirectionsCriteria.ApproachesCriteria].
+         * @return this builder for chaining options together
+         * @since 0.15.0
+         */
+        fun addApproaches(vararg approaches: String?): Builder {
+            directionsBuilder.addApproaches(*approaches)
+            return this
+        }
+
+        /**
+         * Optionally, set which input coordinates should be treated as waypoints / separate legs.
+         * Note: coordinate indices not added here act as silent waypoints
+         *
+         *
+         * Most useful in combination with <tt>steps=true</tt> and requests based on traces
+         * with high sample rates. Can be an index corresponding to any of the input coordinates,
+         * but must contain the first ( 0 ) and last coordinates' index separated by <tt>;</tt>.
+         *
+         * @param indices integer array of coordinate indices to be used as waypoints
+         * @return this builder for chaining options together
+         */
+        fun addWaypointIndices(@IntRange(from = 0) vararg indices: Int): Builder {
+            val result = Array(indices.size) { 0 }
+            var index = 0
+            for (i in indices) {
+                result[index++] = i
+            }
+            directionsBuilder.addWaypointIndices(*result)
+            return this
+        }
+
+        /**
+         * Custom names for waypoints used for the arrival instruction,
+         * each separated by <tt>;</tt>. Values can be any string and total number of all characters cannot
+         * exceed 500. If provided, the list of <tt>waypointNames</tt> must be the same length as the list of
+         * coordinates, but you can skip a coordinate and show its position with the <tt>;</tt> separator.
+         *
+         * @param waypointNames Custom names for waypoints used for the arrival instruction.
+         * @return this builder for chaining options together
+         * @since 0.15.0
+         */
+        fun addWaypointNames(vararg waypointNames: String): Builder {
+            directionsBuilder.addWaypointNames(*waypointNames)
+            return this
+        }
+
+        /**
+         * A list of coordinate pairs used to specify drop-off
+         * locations that are distinct from the locations specified in coordinates.
+         * If this parameter is provided, the Directions API will compute the side of the street,
+         * <tt>left</tt> or <tt>right</tt>, for each target based on the <tt>waypoint_targets</tt>
+         * and the driving direction.
+         * The <tt>maneuver.modifier</tt>, banner and voice instructions will be updated with the computed
+         * side of street. The number of waypoint targets must be the same as the number of coordinates,
+         * but you can skip a coordinate pair and show its position in the list adding <tt>null</tt>.
+         * Must be used with <tt>steps=true</tt>.
+         *
+         * @param waypointTargets [Point] coordinates for drop-off locations
+         * @return this builder for chaining options together
+         * @since 0.26.0
+         */
+        fun addWaypointTargets(vararg waypointTargets: Point?): Builder {
+            directionsBuilder.addWaypointTargets(*waypointTargets)
             return this
         }
 
@@ -227,7 +598,7 @@ constructor(
          * @param eventListener to set for OkHttp
          * @return this builder for chaining options together
          */
-        internal fun eventListener(eventListener: EventListener): Builder {
+        fun eventListener(eventListener: EventListener): Builder {
             directionsBuilder.eventListener(eventListener)
             return this
         }
@@ -244,6 +615,33 @@ constructor(
         }
 
         /**
+         * Sets allowed direction of travel when departing intermediate waypoints. If true the route
+         * will continue in the same direction of travel. If false the route may continue in the
+         * opposite direction of travel. API defaults to true for
+         * [DirectionsCriteria.PROFILE_DRIVING] and false for
+         * [DirectionsCriteria.PROFILE_WALKING] and [DirectionsCriteria.PROFILE_CYCLING].
+         *
+         * @param continueStraight boolean true if you want to always continue straight, else false.
+         * @return this builder for chaining options together
+         */
+        fun continueStraight(continueStraight: Boolean): Builder {
+            directionsBuilder.continueStraight(continueStraight)
+            return this
+        }
+
+        /**
+         * Sets a [NavigationWalkingOptions] object which contains options for use with the
+         * walking profile.
+         *
+         * @param navigationWalkingOptions object holding walking options
+         * @return this builder for chaining options together
+         */
+        fun walkingOptions(navigationWalkingOptions: NavigationWalkingOptions): Builder {
+            directionsBuilder.walkingOptions(navigationWalkingOptions.walkingOptions)
+            return this
+        }
+
+        /**
          * Optionally create a [Builder] based on all variables
          * from given [RouteOptions].
          *
@@ -256,103 +654,45 @@ constructor(
          * @return this builder for chaining options together
          * @since 0.9.0
          */
-        internal fun routeOptions(options: RouteOptionsNavigation): Builder {
-            options.baseUrl?.let {
-                directionsBuilder.baseUrl(it)
+        fun routeOptions(options: RouteOptions): Builder {
+
+            if (!TextUtils.isEmpty(options.baseUrl())) {
+                directionsBuilder.baseUrl(options.baseUrl())
             }
 
-            options.user?.let {
-                directionsBuilder.user(it)
+            if (!TextUtils.isEmpty(options.language())) {
+                directionsBuilder.language(java.util.Locale(options.language()))
             }
 
-            options.profile?.let {
-                directionsBuilder.profile(it)
+            if (options.alternatives() != null) {
+                directionsBuilder.alternatives(options.alternatives())
             }
 
-            origin = options.coordinates.first()
-
-            waypoints.addAll(options.coordinates.toMutableList().also {
-                it.remove(options.coordinates.first())
-                it.remove(options.coordinates.last())
-            })
-
-            destination = options.coordinates.last()
-
-            options.alternatives?.let {
-                directionsBuilder.alternatives(it)
+            if (!TextUtils.isEmpty(options.profile())) {
+                directionsBuilder.profile(options.profile())
             }
 
-            options.language?.let {
-                directionsBuilder.language(Locale(it))
+            if (options.alternatives() != null) {
+                directionsBuilder.alternatives(options.alternatives())
             }
 
-            options.radiuses?.let { radiuses ->
-                if (radiuses.isNotEmpty()) {
-                    val result =
-                        // TODO Convert from String separated by ; into an array of double
-                        radiuses.split(SEMICOLON.toRegex()).dropLastWhile { it.isEmpty() } as DoubleArray
-                    directionsBuilder.radiuses(*result)
-                }
+            if (!TextUtils.isEmpty(options.voiceUnits())) {
+                directionsBuilder.voiceUnits(options.voiceUnits())
             }
 
-            options.bearings?.let { bearings ->
-                if (bearings.isNotEmpty()) {
-                    // TODO Convert from String separated by ; into pairs of angle and tolerance
-                    bearings.split(SEMICOLON.toRegex()).dropLastWhile { it.isEmpty() }.forEach {
-                        directionsBuilder.addBearing(it.toDouble(), it.toDouble())
-                    }
-                }
+            if (!TextUtils.isEmpty(options.user())) {
+                directionsBuilder.user(options.user())
             }
 
-            options.continueStraight?.let {
-                directionsBuilder.continueStraight(it)
+            if (!TextUtils.isEmpty(options.accessToken())) {
+                directionsBuilder.accessToken(options.accessToken())
             }
 
-            options.roundaboutExits?.let {
-                directionsBuilder.roundaboutExits(it)
+            if (!TextUtils.isEmpty(options.annotations())) {
+                directionsBuilder.annotations(options.annotations())
             }
 
-            options.geometries?.let {
-                directionsBuilder.geometries(it)
-            }
-
-            options.overview?.let {
-                directionsBuilder.overview(it)
-            }
-
-            options.steps?.let {
-                directionsBuilder.steps(it)
-            }
-
-            options.annotations?.let {
-                directionsBuilder.annotations(it)
-            }
-
-            options.voiceInstructions?.let {
-                directionsBuilder.voiceInstructions(it)
-            }
-
-            options.bannerInstructions?.let {
-                directionsBuilder.bannerInstructions(it)
-            }
-
-            options.voiceUnits?.let {
-                directionsBuilder.voiceUnits(it)
-            }
-
-            options.accessToken?.let {
-                directionsBuilder.accessToken(it)
-            }
-
-            options.requestUuid?.let {
-                // TODO Check if needed as it is only set at response time
-            }
-
-            options.exclude?.let {
-                directionsBuilder.exclude(it)
-            }
-
-            options.approaches?.let { approaches ->
+            options.approaches()?.let { approaches ->
                 if (approaches.isNotEmpty()) {
                     val result =
                         approaches.split(SEMICOLON.toRegex()).dropLastWhile { it.isEmpty() }
@@ -361,14 +701,14 @@ constructor(
                 }
             }
 
-            options.waypointIndices?.let { waypointIndices ->
+            options.waypointIndices()?.let { waypointIndices ->
                 if (waypointIndices.isNotEmpty()) {
                     val splitWaypointIndices = parseWaypointIndices(waypointIndices)
                     directionsBuilder.addWaypointIndices(*splitWaypointIndices)
                 }
             }
 
-            options.waypointNames?.let { waypointNames ->
+            options.waypointNames()?.let { waypointNames ->
                 if (waypointNames.isNotEmpty()) {
                     val names =
                         waypointNames.split(SEMICOLON.toRegex()).dropLastWhile { it.isEmpty() }
@@ -377,15 +717,16 @@ constructor(
                 }
             }
 
-            options.waypointTargets?.let { waypointTargets ->
+            options.waypointTargets()?.let { waypointTargets ->
                 if (waypointTargets.isNotEmpty()) {
                     val splitWaypointTargets = parseWaypointTargets(waypointTargets)
                     directionsBuilder.addWaypointTargets(*splitWaypointTargets)
                 }
             }
 
-            options.walkingOptions?.let {
-                directionsBuilder.walkingOptions(it.mapToWalkingOptions())
+            val walkingOptions = options.walkingOptions()
+            if (walkingOptions != null) {
+                directionsBuilder.walkingOptions(walkingOptions)
             }
 
             return this
@@ -402,7 +743,14 @@ constructor(
             // Set the default values which the user cannot alter.
             assembleWaypoints()
             directionsBuilder
+                .steps(true)
+                .geometries(DirectionsCriteria.GEOMETRY_POLYLINE6)
+                .overview(DirectionsCriteria.OVERVIEW_FULL)
+                .voiceInstructions(true)
+                .bannerInstructions(true)
+                .roundaboutExits(true)
                 .eventListener(eventListener)
+                .enableRefresh(true)
             return NavigationRoute(directionsBuilder.build())
         }
