+++ libnavigation-base/build.gradle
@@ -16,7 +16,6 @@ android {
 
 dependencies {
     api dependenciesList.mapboxCore
-    api dependenciesList.mapboxSdkServices
 
     implementation dependenciesList.supportAnnotation
