+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/util/JsonMapper.kt
@@ -1,0 +1,5 @@
+package com.mapbox.navigation.base.util
+
+interface JsonMapper {
+    fun <T> toJson(obj: T): String
+}
