+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/MapboxMetricsReporter.kt
@@ -6,6 +6,7 @@ import com.mapbox.android.telemetry.MapboxTelemetry
 import com.mapbox.navigation.base.metrics.MetricEvent
 import com.mapbox.navigation.base.metrics.MetricsObserver
 import com.mapbox.navigation.base.metrics.MetricsReporter
+import com.mapbox.navigation.base.util.JsonMapper
 import com.mapbox.navigation.metrics.internal.utils.extensions.toTelemetryEvent
 import com.mapbox.navigation.utils.thread.WorkThreadHandler
 
@@ -69,7 +70,9 @@ object MapboxMetricsReporter : MetricsReporter {
         }
 
         threadWorker.post {
-            metricsObserver?.onMetricUpdated(metricEvent.metric, metricEvent.toJson(gson))
+            metricsObserver?.onMetricUpdated(metricEvent.metric, metricEvent.toJson(object : JsonMapper {
+                override fun <T> toJson(obj: T): String = gson.toJson(obj)
+            }))
         }
     }
