+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/internal/NavigationAppUserTurnstileEvent.kt
@@ -1,9 +1,9 @@
 package com.mapbox.navigation.metrics.internal
 
-import com.google.gson.Gson
 import com.mapbox.android.telemetry.AppUserTurnstile
 import com.mapbox.navigation.base.internal.metrics.NavigationMetrics
 import com.mapbox.navigation.base.metrics.MetricEvent
+import com.mapbox.navigation.base.util.JsonMapper
 
 internal class NavigationAppUserTurnstileEvent(
     val event: AppUserTurnstile
@@ -12,5 +12,5 @@ internal class NavigationAppUserTurnstileEvent(
     override val metric: String
         get() = NavigationMetrics.APP_USER_TURNSTILE
 
-    override fun toJson(gson: Gson): String = gson.toJson(this)
+    override fun toJson(jsonMapper: JsonMapper): String = jsonMapper.toJson(this)
 }
